调通ListAction 使用parameterClass="java.util.Map"实现多参数传递

+++ trunk/Rax/src/rax/dao/ArticleCategoryDao.java
@@ -7,9 +7,9 @@ import rax.model.ArticleCategory;
 public interface ArticleCategoryDao extends GenericDao<ArticleCategory, Long> {
 
     @Override
-    public long count();
+    public int count();
 
-    public long countByCategoryId(Long id);
+    public int countByCategoryId(Long id);
 
     @Override
     public Long create(ArticleCategory newInstance);
